detector/requirements.txt
Flask==3.0.3
flask_caching==2.3.0
flask_cors==5.0.0
flask_limiter==3.8.0
flask_restx==1.3.0
python-dotenv==1.0.1
joblib==1.4.2
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
python-socketio==5.11.4
scipy==1.14.1
setuptools==75.1.0
pyWavelets==1.7.0
scikit-learn==1.5.2
lightgbm==4.5.0
xgboost==2.1.1
requests==2.26.0